Add createdBy parameter to `/proposals`
This adds the ability to specify a createdBy filter when loading
proposals.  It also improves the way our `loadProposal` operation
and query work so that the authorization context is used in a more
complete way.

Issue #928 Add ability to filter by creator
Issue #927 Allow administrative users to view all data regardless of ownership