Rename basefield label and description
This commit renames the label and description field on the basefield
schema to 'default_label' and 'default_description' in postgres, and
accommodates the changes in Typescript. Whereas previously these fields
were the canonical definitions of a basefield, they are now meant to
represent the default semantic meaning of a basefield, in the context of
providing multiple localizations of a basefield. Moving forward, when a
basefield is served to a user, we want to be able to provide them with
their preferred language, but if the preferred language is not available
as a localization, we want to provide them with the default label and
description.